+++ containers/daap-python-base/tests/unit/versioning_test.py
@@ -7,7 +7,8 @@ from unittest.mock import patch
 
 import pytest
 from botocore.exceptions import ClientError
-from versioning import InvalidUpdate, VersionCreator
+from data_product_metadata import DataProductSchema
+from versioning import InvalidUpdate, VersionManager
 
 test_metadata = {
     "name": "test_product",
@@ -228,7 +229,7 @@ minor_inputs = [
 ]
 
 
-class TestVersionCreator:
+class TestVersionManager:
     @pytest.fixture(autouse=True)
     def setup(self, metadata_bucket, s3_client, data_product_name):
         self.s3_client = s3_client
@@ -258,9 +259,9 @@ class TestVersionCreator:
         input_data = dict(**test_metadata)
         input_data["description"] = "New description"
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
 
-        version = version_creator.update_metadata(input_data)
+        version = version_manager.update_metadata(input_data)
 
         assert version == "v1.1"
         self.assert_has_keys({"test_product/v1.1/metadata.json"}, "v1.1")
@@ -273,9 +274,9 @@ class TestVersionCreator:
             Key="test_product/v1.0/test_table/schema.json",
         )
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
 
-        version, changes = version_creator.update_schema(input_data, "test_table")
+        version, changes = version_manager.update_schema(input_data, "test_table")
 
         assert version == "v1.1"
         assert changes == expected
@@ -287,35 +288,44 @@ class TestVersionCreator:
             "v1.1",
         )
 
-    def test_create_schema_version_new(self):
+    def test_create_schema_version_new(self, s3_client, table_name):
         data_product_name = test_metadata["name"]
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
-
-        version = version_creator.create_schema(
-            input_data=test_metadata, table_name="test_table"
+        version_manager = VersionManager(data_product_name, logging.getLogger())
+        schema = DataProductSchema(
+            data_product_name=data_product_name,
+            table_name=table_name,
+            logger=logging.getLogger(),
+            input_data=test_schema,
         )
+        version = version_manager.create_schema(schema)
 
         assert version == "v1.0"
         self.assert_has_keys(
             {
                 "test_product/v1.0/metadata.json",
+                "test_product/v1.0/table-name/schema.json",
             },
             "v1.0",
             data_product_name,
         )
 
-    def test_create_schema_version_exists_bump(self):
+    def test_create_schema_version_exists_bump(self, table_name):
         data_product_name = test_metadata_with_schemas["name"]
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
 
-        version = version_creator.create_schema(
-            input_data=test_metadata_with_schemas, table_name="test_table"
+        schema = DataProductSchema(
+            data_product_name="test_product_with_schemas",
+            table_name=table_name,
+            logger=logging.getLogger(),
+            input_data=test_schema,
         )
+        version = version_manager.create_schema(schema)
 
         assert version == "v1.1"
         self.assert_has_keys(
             {
                 "test_product_with_schemas/v1.1/metadata.json",
+                "test_product_with_schemas/v1.1/table-name/schema.json",
             },
             "v1.1",
             data_product_name,
@@ -329,9 +339,9 @@ class TestVersionCreator:
             Key="test_product/v1.0/test_table/schema.json",
         )
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
 
-        version, changes = version_creator.update_schema(input_data, "test_table")
+        version, changes = version_manager.update_schema(input_data, "test_table")
 
         assert version == "v2.0"
         assert changes == expected
@@ -351,9 +361,9 @@ class TestVersionCreator:
         )
         input_data = copy.deepcopy(test_schema)
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
         with pytest.raises(InvalidUpdate):
-            version, changes = version_creator.update_schema(input_data, "test_table")
+            version, changes = version_manager.update_schema(input_data, "test_table")
 
     def test_copies_schemas(self, s3_client):
         s3_client.put_object(
@@ -365,9 +375,9 @@ class TestVersionCreator:
         input_data = dict(**test_metadata)
         input_data["description"] = "New description"
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
 
-        version_creator.update_metadata(input_data)
+        version_manager.update_metadata(input_data)
 
         self.assert_has_keys(
             {
@@ -381,19 +391,19 @@ class TestVersionCreator:
         input_data = dict(**test_metadata)
         input_data["name"] = "new name"
 
-        version_creator = VersionCreator(test_metadata["name"], logging.getLogger())
+        version_manager = VersionManager(test_metadata["name"], logging.getLogger())
 
         with pytest.raises(InvalidUpdate):
-            version_creator.update_metadata(input_data)
+            version_manager.update_metadata(input_data)
 
     def test_cannot_update_product_that_does_not_exist(self):
         input_data = dict(**test_metadata)
         input_data["description"] = "New description"
 
-        version_creator = VersionCreator("does_not_exist", logging.getLogger())
+        version_manager = VersionManager("does_not_exist", logging.getLogger())
 
         with pytest.raises(InvalidUpdate):
-            version_creator.update_metadata(input_data)
+            version_manager.update_metadata(input_data)
 
 
 class TestUpdateMetadataRemoveSchema:
@@ -428,10 +438,10 @@ class TestUpdateMetadataRemoveSchema:
     def test_success(
         self, s3_client, create_raw_and_curated_data, data_product_name, glue_client
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0"]
         with patch("glue_utils.glue_client", glue_client):
-            version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+            version_manager.update_metadata_remove_schemas(schema_list=schema_list)
             schema_prefix = f"{data_product_name}/v2.0/metadata.json"
             response = s3_client.list_objects_v2(
                 Bucket=self.bucket_name,
@@ -440,11 +450,11 @@ class TestUpdateMetadataRemoveSchema:
         assert response.get("KeyCount") == 1
 
     def test_invalid_schemas(self, data_product_name):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema3", "schema4"]
 
         with pytest.raises(InvalidUpdate) as exc:
-            version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+            version_manager.update_metadata_remove_schemas(schema_list=schema_list)
         assert (
             str(exc.value)
             == "Invalid schemas found in schema_list: ['schema3', 'schema4']"
@@ -458,11 +468,11 @@ class TestUpdateMetadataRemoveSchema:
         glue_client,
         table_name,
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0", "schema1"]
         with patch("glue_utils.glue_client", glue_client):
             with pytest.raises(ValueError) as exc:
-                version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+                version_manager.update_metadata_remove_schemas(schema_list=schema_list)
                 assert (
                     str(exc.value)
                     == f"Could not locate glue table '{table_name}' in database '{data_product_name}'"
@@ -471,7 +481,7 @@ class TestUpdateMetadataRemoveSchema:
     def test_schema_glue_table_deleted(
         self, s3_client, create_raw_and_curated_data, data_product_name, glue_client
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0"]
         with patch("glue_utils.glue_client", glue_client):
             table = glue_client.get_table(
@@ -482,7 +492,7 @@ class TestUpdateMetadataRemoveSchema:
 
             with pytest.raises(ClientError) as exc:
                 # Call the handler
-                version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+                version_manager.update_metadata_remove_schemas(schema_list=schema_list)
 
                 table = glue_client.get_table(
                     DatabaseName=data_product_name, Name=f"{schema_list[0]})"
@@ -498,7 +508,7 @@ class TestUpdateMetadataRemoveSchema:
         glue_client,
         data_product_versions,
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0"]
         with patch("glue_utils.glue_client", glue_client):
             # Validate we have the required number of files
@@ -518,7 +528,7 @@ class TestUpdateMetadataRemoveSchema:
                 assert response.get("KeyCount") == 10
 
             # Call the handler
-            version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+            version_manager.update_metadata_remove_schemas(schema_list=schema_list)
 
             # Validate files are deleted
             for version in data_product_versions:
@@ -544,12 +554,12 @@ class TestUpdateMetadataRemoveSchema:
         glue_client,
         data_product_versions,
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0"]
 
         with patch("glue_utils.glue_client", glue_client):
             # Call the handler
-            version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+            version_manager.update_metadata_remove_schemas(schema_list=schema_list)
             schema_prefix = f"{data_product_name}/v2.0/{schema_list[0]}/schema.json"
             response = s3_client.list_objects_v2(
                 Bucket=self.bucket_name,
@@ -565,12 +575,12 @@ class TestUpdateMetadataRemoveSchema:
         glue_client,
         data_product_versions,
     ):
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
         schema_list = ["schema0"]
 
         with patch("glue_utils.glue_client", glue_client):
             # Call the handler
-            version_creator.update_metadata_remove_schemas(schema_list=schema_list)
+            version_manager.update_metadata_remove_schemas(schema_list=schema_list)
             for i in range(1, 3):
                 schema_prefix = f"{data_product_name}/v2.0/schema{i}/schema.json"
                 response = s3_client.list_objects_v2(
@@ -642,11 +652,11 @@ class TestRemoveAllVersions:
         assert response.get("KeyCount") == 30
 
         #######################################################################
-        version_creator = VersionCreator(data_product_name, logging.getLogger())
+        version_manager = VersionManager(data_product_name, logging.getLogger())
 
         with patch("glue_utils.glue_client", glue_client):
             # Call the handler
-            version_creator.remove_all_versions_of_data_product()
+            version_manager.remove_all_versions_of_data_product()
 
             # Assert that all metadata files have been deleted
             prefix = f"{data_product_name}/"
@@ -688,10 +698,10 @@ class TestRemoveAllVersions:
         data_product_name,
         glue_client,
     ):
-        version_creator = VersionCreator("fake_data_product_name", logging.getLogger())
+        version_manager = VersionManager("fake_data_product_name", logging.getLogger())
         with pytest.raises(ValueError) as exc:
             # Call the handler
-            version_creator.remove_all_versions_of_data_product()
+            version_manager.remove_all_versions_of_data_product()
         assert (
             str(exc.value) == "Could not locate glue database 'fake_data_product_name'"
         )