+++ containers/daap-delete-table-for-data-product/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.1] - 2023-11-16
+
+### Changed
+
+- renamed VersionCreator to VersionManager to match versioning module
+
 ## [2.1.0] - 2023-11-13
 
 ### Changed