+++ containers/daap-update-metadata/CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2023-11-16
+
+### Changed
+
+- renamed VersionCreator to VersionManager to match versioning module
+
 ## [1.1.0] - 2023-11-13
 
 - update base image to 6.2.0